Pull out linux_host into a library.
This commit moves all the linux_host code into a library and exposes the
Main function. This allows callers to create linux_host binaries that
depend on different Tao implements, added using the tao.Register
function.
